Add a failing test case for the XHProf PHP 5.5 segfault
Summary:
This adds a failing test case for the bug discussed in <https://github.com/phacility/xhprof/pull/33>.

It also adds a `bin/xhprofile` script which makes it easier to test stuff like this, by invoking the profiler on some other script. This isn't hugely useful in production but is valuable diagnostically, and helped me reduce this test case.

Test Plan:
  - Ran `arc unit` and got a failure.
  - Verified that this test passes if `class_exists()` is commented out (which makes the test not segfault).

Reviewers: btrahan

Reviewed By: btrahan

Subscribers: epriestley

Differential Revision: https://secure.phabricator.com/D10374